+++ src/Options/Definitions.js
@@ -11,6 +11,11 @@ module.exports.ParseServerOptions = {
     help: 'account lockout policy for failed login attempts',
     action: parsers.objectParser,
   },
+  advisoryChecks: {
+    env: 'PARSE_SERVER_ADVISORY_CHECKS',
+    help: 'view recommendations for server improvements',
+    default: false,
+  },
   allowClientClassCreation: {
     env: 'PARSE_SERVER_ALLOW_CLIENT_CLASS_CREATION',
     help: 'Enable (or disable) client class creation, defaults to true',
